Solve log level and rename handler
