+++ lib/logger/lib/logger/handler.ex
@@ -1,22 +1,8 @@
-defmodule Logger.LegacyHandler do
+defmodule Logger.Handler do
   @moduledoc false
   alias Logger.Counter
 
   @internal_keys [:counter]
-
-  ## API used by other Logger modules
-
-  def load_log_level do
-    # TODO: Deprecate if level is warning.
-    # PENDING: Fallback to kernel.logger_level if logger level is not set.
-    level =
-      case Application.fetch_env!(:logger, :level) do
-        :warn -> :warning
-        level -> level
-      end
-
-    :ok = :logger.set_primary_config(:level, level)
-  end
 
   ## Config management
