+++ lib/logger/test/logger/legacy_handler_test.exs
@@ -1,4 +1,4 @@
-defmodule Logger.LegacyHandlerTest do
+defmodule Logger.HandlerTest do
   use Logger.Case
   @moduletag :logger
